+++ src/jvm/storm/kafka/FailedFetchException.java
@@ -1,6 +1,11 @@
-package storm.kafka.trident;
+package storm.kafka;
 
 public class FailedFetchException extends RuntimeException {
+
+    public FailedFetchException(String message) {
+        super(message);
+    }
+
     public FailedFetchException(Exception e) {
         super(e);
     }