+++ src/jvm/storm/kafka/KafkaUtils.java
@@ -1,20 +1,29 @@
-package storm.kafka.trident;
+package storm.kafka;
 
 import backtype.storm.metric.api.IMetric;
+import kafka.api.FetchRequest;
+import kafka.api.FetchRequestBuilder;
 import kafka.api.PartitionOffsetRequestInfo;
 import kafka.common.TopicAndPartition;
+import kafka.javaapi.FetchResponse;
 import kafka.javaapi.OffsetRequest;
 import kafka.javaapi.consumer.SimpleConsumer;
+import kafka.javaapi.message.ByteBufferMessageSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import storm.kafka.*;
+import storm.kafka.trident.IBrokerReader;
+import storm.kafka.trident.StaticBrokerReader;
+import storm.kafka.trident.ZkBrokerReader;
 
+import java.net.ConnectException;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
+
 public class KafkaUtils {
+
     public static final Logger LOG = LoggerFactory.getLogger(KafkaUtils.class);
     private static final int NO_OFFSET = -5;
 
@@ -108,5 +117,43 @@ public class KafkaUtils {
                 }
             }
         }
+    }
+
+    public static ByteBufferMessageSet fetchMessages(KafkaConfig config, SimpleConsumer consumer, Partition partition, long offset) {
+        ByteBufferMessageSet msgs = null;
+        String topic = config.topic;
+        int partitionId = partition.partition;
+        for (int errors = 0; errors < 2 && msgs == null; errors++) {
+            FetchRequestBuilder builder = new FetchRequestBuilder();
+            FetchRequest fetchRequest = builder.addFetch(topic, partitionId, offset, config.fetchSizeBytes).
+                    clientId(config.clientId).build();
+            FetchResponse fetchResponse;
+            try {
+                fetchResponse = consumer.fetch(fetchRequest);
+            } catch (Exception e) {
+                if (e instanceof ConnectException) {
+                    throw new FailedFetchException(e);
+                } else {
+                    throw new RuntimeException(e);
+                }
+            }
+            if (fetchResponse.hasError()) {
+                KafkaError error = KafkaError.getError(fetchResponse.errorCode(topic, partitionId));
+                if (error.equals(KafkaError.OFFSET_OUT_OF_RANGE) && config.useStartOffsetTimeIfOffsetOutOfRange && errors == 0) {
+                    long startOffset = getOffset(consumer, topic, partitionId, config.startOffsetTime);
+                    LOG.warn("Got fetch request with offset out of range: [" + offset + "]; " +
+                            "retrying with default start offset time from configuration. " +
+                            "configured start offset time: [" + config.startOffsetTime + "] offset: [" + startOffset + "]");
+                    offset = startOffset;
+                } else {
+                    String message = "Error fetching data from [" + partition + "] for topic [" + topic + "]: [" + error + "]";
+                    LOG.error(message);
+                    throw new FailedFetchException(message);
+                }
+            } else {
+                msgs = fetchResponse.messageSet(topic, partitionId);
+            }
+        }
+        return msgs;
     }
 }