+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 0.3.0
 * updated partition path in zookeeper
+* added error handling for fetch request